+++ nvim/highlights.scm
@@ -51,7 +51,6 @@
   "key"
   "new"
   "not"
-  "null"
   "of"
   "on"
   "select"